+++ src/Controller/RegisterController.php
@@ -13,6 +13,7 @@ use Ingenerator\Warden\Core\Support\InteractorRequestFactory;
 use Ingenerator\Warden\Core\Support\UrlProvider;
 use Ingenerator\Warden\Core\UserSession\UserSession;
 use Ingenerator\Warden\UI\Kohana\Form\Fieldset;
+use Ingenerator\Warden\UI\Kohana\Message\Register\InvalidRegisterLinkMessage;
 use Ingenerator\Warden\UI\Kohana\Message\Register\RegistrationSuccessMessage;
 use Ingenerator\Warden\UI\Kohana\View\RegistrationView;
 
@@ -48,27 +49,52 @@ class RegisterController extends WardenBaseController
         $this->register_interactor = $register_interactor;
         $this->register_view       = $register_view;
         $this->session             = $session;
-        $this->urls = $urls;
+        $this->urls                = $urls;
     }
 
     public function before()
     {
         parent::before();
         $this->redirectHomeIfLoggedIn($this->session, $this->urls);
-
-        // if configured to require email confirm before registration and there is no token param
-        if ( ! $this->request->query('token')) {
-            $this->redirect($this->urls->getRegisterVerifyEmailUrl($this->request->query('email')));
-        }
     }
 
     public function action_get()
     {
-        $data = [
-            'email'                    => $this->request->query('email'),
-            'email_confirmation_token' => $this->request->query('token'),
-        ];
-        $this->displayRegistrationForm(new Fieldset($data, []));
+        $request     = $this->makeRegistrationRequest();
+        $token_state = $this->register_interactor->validateToken($request);
+        if ( ! $token_state->isValid()) {
+            $this->handleInvalidLinkFailure($request->getEmail());
+        }
+
+        $this->displayRegistrationForm(
+            new Fieldset(
+                [
+                    'email'                    => $request->getEmail(),
+                    'email_confirmation_token' => $request->getToken()
+                ], []
+            )
+        );
+    }
+
+    /**
+     * @return \Ingenerator\Warden\Core\Interactor\UserRegistrationRequest
+     */
+    protected function makeRegistrationRequest()
+    {
+        if ($this->request->method() === \Request::POST) {
+            $values = $this->request->post();
+        } else {
+            $values = [
+                'email'                    => $this->request->query('email'),
+                'email_confirmation_token' => $this->request->query('token'),
+            ];
+        }
+
+        return $this->makeInteractorRequest(
+            'user_registration',
+            'fromArray',
+            $values
+        );
     }
 
     protected function displayRegistrationForm(Fieldset $fieldset)
@@ -112,7 +138,9 @@ class RegisterController extends WardenBaseController
     protected function handleActiveUserRegistration(User $user)
     {
         if ( ! $this->session->isAuthenticated()) {
-            throw new \UnexpectedValueException('Active user was not authenticated after registration');
+            throw new \UnexpectedValueException(
+                'Active user was not authenticated after registration'
+            );
         }
 
         $this->getPigeonhole()->add(new RegistrationSuccessMessage());
@@ -124,12 +152,27 @@ class RegisterController extends WardenBaseController
         if ($result->isFailureCode(UserRegistrationResponse::ERROR_ALREADY_REGISTERED)) {
             $this->handleRegisterAttemptForExistingUser($this->urls, $result->getEmail());
 
+        } elseif ($result->isFailureCode(
+            UserRegistrationResponse::ERROR_EMAIL_CONFIRMATION_INVALID
+        )) {
+            $this->handleInvalidLinkFailure($result->getEmail());
+
         } elseif ($result->isFailureCode(UserRegistrationResponse::ERROR_DETAILS_INVALID)) {
-            $this->displayRegistrationForm(new Fieldset($this->request->post(), $result->getValidationErrors()));
+            $this->displayRegistrationForm(
+                new Fieldset($this->request->post(), $result->getValidationErrors())
+            );
 
         } else {
-            throw new \UnexpectedValueException('Unexpected registration failure: '.$result->getFailureCode());
+            throw new \UnexpectedValueException(
+                'Unexpected registration failure: '.$result->getFailureCode()
+            );
         }
+    }
+
+    protected function handleInvalidLinkFailure($email)
+    {
+        $this->getPigeonhole()->add(new InvalidRegisterLinkMessage);
+        $this->redirect($this->urls->getRegisterVerifyEmailUrl($email));
     }
 
 }