+++ test/unit/View/AbstractFormViewTest.php
@@ -18,14 +18,13 @@ abstract class AbstractFormViewTest extends \PHPUnit_Framework_TestCase
      */
     public function test_it_throws_if_fields_is_not_a_fieldset()
     {
-        $this->newSubject()->display(['fields' => 'random stuff']);
+        $this->newSubjectDisplaying(['fields' => 'random stuff']);
     }
 
     public function test_it_exposes_fields_variable()
     {
-        $fields = new Fieldset([], []);
-        $subject = $this->newSubject();
-        $subject->display(['fields' => $fields]);
+        $fields  = new Fieldset([], []);
+        $subject = $this->newSubjectDisplaying(['fields' => $fields]);
         $this->assertSame($fields, $subject->fields);
     }
 
@@ -33,5 +32,18 @@ abstract class AbstractFormViewTest extends \PHPUnit_Framework_TestCase
      * @return AbstractViewModel
      */
     abstract protected function newSubject();
+
+    /**
+     * @param array $custom_vars
+     *
+     * @return AbstractViewModel
+     */
+    protected function newSubjectDisplaying(array $custom_vars)
+    {
+        $vars    = array_merge(['fields' => new Fieldset([], [])], $custom_vars);
+        $subject = $this->newSubject();
+        $subject->display($vars);
+        return $subject;
+    }
 
 }